Put sync for create mysql pos file.
The sync command is nescessary because we do a snapshot before
create a file. the new file dosent appear on snapshot because
the SO buffePut sync for create mysql pos file.

The sync command is nescessary because we do a snapshot before
create a file. the new file dosent appear on snapshot because
the SO bufferr